requirements.txt
asgiref==3.4.1
astroid==2.4.2
attrs==21.2.0
autobahn==21.11.1
autograd @ git+https://github.com/HIPS/autograd@01eacff7a4f12e6f7aebde7c4cb4c1c2633f217d
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==2.2.17
django-rest-framework==0.1.0
djangorestframework==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pastebin==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylint==2.5.3
pyOpenSSL==21.0.0
pytz==2021.3
rope==0.18.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
wincertstore==0.2
wrapt==1.12.1
zope.interface==5.4.0
